Renaming the app template.
