+++ app/build.gradle
@@ -7,7 +7,7 @@ android {
     compileSdk 31
 
     defaultConfig {
-        applicationId "com.adammcneilly.TODOREPLACE"
+        applicationId "com.adammcneilly.pocketleague"
         minSdk 21
         targetSdk 31
         versionCode 1
