+++ app/src/main/java/com/pocketleague/theme/Color.kt
@@ -1,6 +1,6 @@
 @file:Suppress("MagicNumber")
 
-package com.template.theme
+package com.pocketleague.theme
 
 import androidx.compose.ui.graphics.Color
