+++ app/src/main/java/com/pocketleague/theme/Shape.kt
@@ -1,4 +1,4 @@
-package com.template.theme
+package com.pocketleague.theme
 
 import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.material.Shapes
