+++ app/src/main/java/com/pocketleague/theme/Theme.kt
@@ -1,4 +1,4 @@
-package com.template.theme
+package com.pocketleague.theme
 
 import androidx.compose.foundation.isSystemInDarkTheme
 import androidx.compose.material.MaterialTheme
@@ -28,7 +28,7 @@ private val LightColorPalette = lightColors(
 )
 
 @Composable
-fun TemplateTheme(
+fun PocketLeagueTheme(
     darkTheme: Boolean = isSystemInDarkTheme(),
     content: @Composable() () -> Unit
 ) {
