+++ app/src/main/java/com/pocketleague/theme/Type.kt
@@ -1,4 +1,4 @@
-package com.template.theme
+package com.pocketleague.theme
 
 import androidx.compose.material.Typography
 import androidx.compose.ui.text.TextStyle
